server/requirements.txt
cairocffi==1.0.2
CairoSVG==2.7.0
cffi==1.12.3
Flask==1.1.1
gunicorn==19.9.0
requests==2.24.0
Flask-Limiter==1.3.1
python-dotenv==0.14.0
WeasyPrint==51
